requirements.txt
gradio
transformers
diffusers
ctransformers
Pillow
streamlit
accelerate
pypdf
peft
nltk
newspaper3k
lxml_html_clean
requests
bs4